Refactor to use barrel for eovim folder
